Refactor: create refactor folder; split files
