+++ scanapi/__init__.py
@@ -3,7 +3,7 @@ name = "scanapi"
 import click
 import logging
 
-from scanapi.tree import EndpointNode
+from scanapi.refactor.tree import EndpointNode
 from scanapi.reporter import Reporter
 from scanapi.settings import SETTINGS
 from scanapi.yaml_loader import load_yaml
