+++ scanapi/refactor/evaluators/__init__.py
@@ -1,0 +1,3 @@
+from scanapi.refactor.evaluators.string_evaluator import StringEvaluator
+
+__all__ = ["StringEvaluator"]
