+++ scanapi/refactor/tree/__init__.py
@@ -1,0 +1,4 @@
+from scanapi.refactor.tree.endpoint_node import EndpointNode
+from scanapi.refactor.tree.request_node import RequestNode
+
+__all__ = ["EndpointNode", "RequestNode"]
