+++ tests/unit/refactor/tree/test_endpoint_node.py
@@ -1,8 +1,6 @@
-import os
 import pytest
 
-from scanapi.errors import BadConfigurationError
-from scanapi.tree import EndpointNode, RequestNode, StringEvaluator
+from scanapi.refactor.tree import EndpointNode
 
 
 class TestEndpointNode:
@@ -47,6 +45,15 @@ class TestEndpointNode:
             assert len(requests) == 2
 
         class TestPath:
+            @pytest.fixture
+            def mock_evaluate(self, mocker):
+                mock_func = mocker.patch(
+                    "scanapi.refactor.tree.endpoint_node.StringEvaluator.evaluate"
+                )
+                mock_func.return_value = ""
+
+                return mock_func
+
             def test_when_parent_has_no_url(self):
                 base_path = "http://foo.com"
                 node = EndpointNode({"path": base_path}, parent=EndpointNode({}))
@@ -56,12 +63,20 @@ class TestEndpointNode:
                 base_path = "http://foo.com/api"
                 parent = EndpointNode({"path": base_path})
                 node = EndpointNode({"path": "/foo"}, parent=parent)
-                assert node.path == f"{base_path}/foo"
+                assert node.path == f"http://foo.com/api/foo"
 
             def test_with_trailing_slashes(self):
                 parent = EndpointNode({"path": "http://foo.com/"})
                 node = EndpointNode({"path": "/foo/"}, parent=parent)
                 assert node.path == "http://foo.com/foo/"
+
+            def test_calls_evaluate(self, mocker, mock_evaluate):
+                parent = EndpointNode({"path": "http://foo.com/"})
+                node = EndpointNode({"path": "/foo/"}, parent=parent)
+                node.path
+                calls = [mocker.call("http://foo.com/"), mocker.call("/foo/")]
+
+                mock_evaluate.assert_has_calls(calls)
 
         class TestHeaders:
             def test_when_parent_has_no_headers(self):
@@ -86,59 +101,3 @@ class TestEndpointNode:
                     parent=EndpointNode({"headers": parent_headers}),
                 )
                 assert node.headers == {"abc": "def", "xxx": "www"}
-
-
-class TestRequestNode:
-    def test_request_with_no_path(self):
-        base_path = "http://foo.com/"
-        request = RequestNode({}, endpoint=EndpointNode({"path": base_path}))
-        assert request.full_url_path == base_path
-
-
-class TestStringEvaluator:
-    class TestEvaluateEnvVar:
-        class TestWhenDoesNotMatchThePattern:
-            test_data = ["no env var", "${var}", "${MyVar}", "${{var}}", "${{VAR}}"]
-
-            @pytest.mark.parametrize("sequence", test_data)
-            def test_should_return_sequence(self, sequence):
-                assert StringEvaluator().evaluate_env_var(sequence) == sequence
-
-        class TestWhenMatchesThePattern:
-            class TestWhenEnvVarIsSetProperly:
-                @pytest.fixture(autouse=True)
-                def base_url_env(self):
-                    os.environ["BASE_URL"] = "https://jsonplaceholder.typicode.com"
-                    os.environ["POST_ID"] = "2"
-
-                test_data = [
-                    ("${BASE_URL}", "https://jsonplaceholder.typicode.com"),
-                    ("${BASE_URL}/posts", "https://jsonplaceholder.typicode.com/posts"),
-                    (
-                        "https://jsonplaceholder.typicode.com/posts/${POST_ID}",
-                        "https://jsonplaceholder.typicode.com/posts/2",
-                    ),
-                    (
-                        "${BASE_URL}/posts/${POST_ID}",
-                        "https://jsonplaceholder.typicode.com/posts/2",
-                    ),
-                ]
-
-                @pytest.mark.parametrize("sequence, expected", test_data)
-                def test_should_return_evaluated_var(self, sequence, expected):
-                    assert StringEvaluator().evaluate_env_var(sequence) == expected
-
-            class TestWhenThereIsNoCorrespondingEnvVar:
-                @pytest.fixture(autouse=True)
-                def remove_base_url_env(self):
-                    if os.environ.get("BASE_URL"):
-                        del os.environ["BASE_URL"]
-
-                def test_should_raise_bad_configuration_error(self):
-                    with pytest.raises(BadConfigurationError) as excinfo:
-                        StringEvaluator().evaluate_env_var("${BASE_URL}")
-
-                    assert (
-                        str(excinfo.value)
-                        == "'BASE_URL' environment variable not set or badly configured"
-                    )
